requirements.txt
python-telegram-bot==20.8
supabase>=2.0.0
python-dotenv
fastapi
uvicorn[standard]
pydantic
Pillow
easyocr
requests